Moves ts files after bob setup
