+++ example/src/config.ts
