+++ src/SessionNativeComponent.ts
