+++ src/VisitableViewNativeComponent.ts
